requirements.txt
python-telegram-bot==21.6
openai==1.51.2
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
dateparser==1.2.0
pydub==0.25.1
pytz==2024.2
pytest==8.0.0
